make file manager selection based on cloud provider config
